Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    double a;
-    scanf("%lf", &a);
-    printf("You entered: %lf", a);
+    float a;
+    scanf("%f", &a);
+    printf("You entered: %f", a);
     return 0;
 }
